+++ evo1/atb/manual.py
@@ -6,5 +6,5 @@ class SeqATBCombatManual(SeqATBCombat):
     def __init__(self, name: str = "Generic", wait_for_battle: bool = False) -> None:
         super().__init__(name=name, wait_for_battle=wait_for_battle)
 
-    def handle_combat(self):
+    def handle_combat(self, should_run: bool = False):
         pass